Simplify tests for pvpc_hourly_pricing
