+++ tests/components/pvpc_hourly_pricing/test_config_flow.py
@@ -2,27 +2,16 @@
 from datetime import datetime
 from unittest.mock import patch
 
-import pytest
 from pytz import timezone
 
 from homeassistant import data_entry_flow
 from homeassistant.components.pvpc_hourly_pricing import ATTR_TARIFF, DOMAIN
 from homeassistant.const import CONF_NAME
 
-from . import FIXTURE_JSON_DATA_2019_03_30, check_valid_state
+from .conftest import check_valid_state
 
-from tests.common import date_util, load_fixture
+from tests.common import date_util
 from tests.test_util.aiohttp import AiohttpClientMocker
-
-
-@pytest.fixture(name="pvpc_aioclient_mock")
-def pvpc_aioclient_mock(aioclient_mock: AiohttpClientMocker):
-    """Create a mock config entry."""
-    aioclient_mock.get(
-        "https://api.esios.ree.es/archives/70/download_json?locale=es&date=2019-03-30",
-        text=load_fixture(f"{DOMAIN}/{FIXTURE_JSON_DATA_2019_03_30}"),
-    )
-    return aioclient_mock
 
 
 async def test_config_flow(hass, pvpc_aioclient_mock: AiohttpClientMocker):
@@ -36,7 +25,7 @@ async def test_config_flow(hass, pvpc_aioclient_mock: AiohttpClientMocker):
     - Check abort stage on name collision
     """
     hass.config.time_zone = timezone("Europe/Madrid")
-    mock_data = {"return_time": datetime(2019, 3, 30, 14, 0, tzinfo=date_util.UTC)}
+    mock_data = {"return_time": datetime(2019, 10, 26, 14, 0, tzinfo=date_util.UTC)}
 
     def mock_now():
         return mock_data["return_time"]
