+++ tests/components/pvpc_hourly_pricing/test_setup.py
@@ -2,26 +2,15 @@
 from datetime import datetime
 from unittest.mock import patch
 
-import pytest
 from pytz import timezone
 
 from homeassistant.components.pvpc_hourly_pricing import ATTR_TARIFF, DOMAIN
 from homeassistant.const import CONF_NAME
 
-from . import FIXTURE_JSON_DATA_2019_03_30, check_valid_state
+from .conftest import check_valid_state
 
-from tests.common import async_setup_component, date_util, load_fixture
+from tests.common import async_setup_component, date_util
 from tests.test_util.aiohttp import AiohttpClientMocker
-
-
-@pytest.fixture
-def pvpc_aioclient_mock(aioclient_mock: AiohttpClientMocker):
-    """Create a mock config entry."""
-    aioclient_mock.get(
-        "https://api.esios.ree.es/archives/70/download_json?locale=es&date=2019-03-30",
-        text=load_fixture(f"{DOMAIN}/{FIXTURE_JSON_DATA_2019_03_30}"),
-    )
-    return aioclient_mock
 
 
 async def _test_setup_from_yaml_config(
@@ -32,7 +21,7 @@ async def _test_setup_from_yaml_config(
     tariff: str = "discrimination",
 ):
     hass.config.time_zone = timezone("Atlantic/Canary")
-    mock_data = {"return_time": datetime(2019, 3, 30, 14, 0, tzinfo=date_util.UTC)}
+    mock_data = {"return_time": datetime(2019, 10, 26, 13, 0, tzinfo=date_util.UTC)}
 
     def mock_now():
         return mock_data["return_time"]
